review fixes

+++ daemon/rpc_pause.go
@@ -9,6 +9,7 @@ import (
 	"github.com/NordSecurity/nordvpn-linux/internal"
 )
 
+// PauseConnection disconnects and schedules a reconnection in a timespan provided in the pause request
 func (r *RPC) PauseConnection(ctx context.Context, in *pb.PauseRequest) (*pb.Payload, error) {
 	if !r.netw.IsVPNActive() {
 		return &pb.Payload{Type: internal.CodeNothingToDo}, nil
@@ -18,7 +19,7 @@ func (r *RPC) PauseConnection(ctx context.Context, in *pb.PauseRequest) (*pb.Pay
 		return &pb.Payload{Type: internal.CodePauseAttemptWhenConnectedToMeshPeer}, nil
 	}
 
-	pauseDuration := time.Duration(in.Seconds * int64(time.Second))
+	pauseDuration := time.Duration(in.Seconds) * time.Second
 	r.pauseManager.ScheduleReconnection(pauseDuration)
 
 	_, err := r.DoDisconnect()
